+++ forui/lib/src/widgets/popover_menu/submenu_item.dart
@@ -200,7 +200,7 @@ class FSubmenuItem extends StatelessWidget with FItemMixin {
     this.submenuCacheExtent,
     this.submenuIntrinsicWidth,
     this.submenuDragStartBehavior = .start,
-    this.submenuSpacing = .zero,
+    this.submenuSpacing = const .spacing(2),
     this.submenuOverflow = .flip,
     this.submenuOffset = .zero,
     this.submenuSemanticsLabel,